Fix patient photo placeholder
The svg patient avatar placeholder is not being rendered in the registration form as it should be. The corresponding placeholder svg
in `esm-patient-attachments/assets/placeholder.svg` is showing up empty on my repo locally as well as on
[GitHub](https://github.com/openmrs/openmrs-esm-patient-chart/edit/master/packages/esm-patient-attachments-app/src/assets/placeholder.svg).

I've added the svg code back into the placeholder svg file and deleted the old png placeholder as it's no longer being used.