MF-886: Program enrollment notification
There should be a subtle way to notify the user that a patient has enrolled in all the available programs.

This PR adds a check that evaluates the available care programs for enrollment eligibility. If a patient has already enrolled in all the available programs, an inline notification is displayed and the Add button (which launches the program enrollment form) gets disabled.