Review feedback
I'll start with the caveat that I haven't quite done Brandon suggested. Now, SWR handles fetching both immunization and concept set data from the backend. I could just as well have moved the logic that determines the configured immunizations and then consolidates them with existing ones and sorts the results to the hook. I haven't done so because this logic only has utility in the ImmunizationsDetailedSummary component - it's probably better off to leave it there. Further, by the time this processing is being data, all the data fetching is complete and this work doesn't depend on fetching any other data. Therefore, it seems fair enough to leave the post-processing to the component.