Fix failing AllergyForm test
This test was failing because of a missing implementation for `useLayoutType`. Despite a manual mock being available for `esm-framework`, this test suite was overriding that manual
mock without providing a suitable alternative (a partial mock or similar). This commit fixes that problem by removing the partial mock of `esm-framework` and relying entirely on the
already available manual mock.