Fix incorrect `useLayoutType` import (#450)
* Fix incorrect useLayoutType import

`useLayoutType` is being imported from the wrong package in the ErrorState component - a change mistakenly introduced in https://github.com/openmrs/openmrs-esm-patient-chart/pull/446.
As a result, an error is displayed instead of the appointments widget. This commit fixes the import and thus fixes the error.

* Fix failing AllergyForm test

This test was failing because of a missing implementation for `useLayoutType`. Despite a manual mock being available for `esm-framework`, this test suite was overriding that manual
mock without providing a suitable alternative (a partial mock or similar). This commit fixes that problem by removing the partial mock of `esm-framework` and relying entirely on the
already available manual mock.