O3-940: Redesign patient avatars
Redesign patient avatars to match the designs outlined in https://app.zeplin.io/project/61434fa756474d5545f65cf4/screen/614362b21dd66d20a276d8b8.

This commit passes the patient name through the state prop into the Display Patient Photo extension. This will be used by the Avatar component to generate the patients initials and display them as a fallback when an image source is not provided. It also gets used as part of the alt text in case there's a problem rendering a user's avatar (so you have something like `John Wilson's avatar` instead of the more generic `Patient avatar`).

`react-avatar` derives the patient's initials and display them as a fallback if an image source is not provided.
   - Set the colour prop to a transparent colour so that the background image (provided by geopattern) can be seen.
   - Set the `photo.imageSrc` value from the aforementioned `usePatientPhoto` hook as the src prop.
   - Finally, set the background image of the avatar to the pattern generated by geopattern using the patient UUID.