(feat): `useSWRImmutable` for fetching vitals concept metadata (#604)
Since v1.0, SWR provides a helper hook useSWRImmutable to mark a resource as immutable. This means that once data is fetched and cached, it will not be requested again. Concept metadata is unlikely to change and as such revalidating is largely unnecessary.

This PR modifies the `useVitalsConceptMetadata` hook to use `useSWRImmutable` for data fetching.