(fix) O3-1220: Vitals header does not update upon form submission
This PR fixes an issue where the vitals header does not auto-update after the vitals and biometrics form gets successfully submitted. This is because the key passed to SWR `mutate` function does not match the URL that the form is configured to POST to.

This PR fixes this issue by using a regular expression to find matching keys and then run `mutate` on them. Because the URL bound to the `useVitals` hook changes based on whether biometrics concepts are included or not, using a RegExp to perform a partial match is a more robust approach than using a fully formed URL that's more likely to not match.