(test) Fix warnings raised when running tests (#637)
* (fix) Fix warnings raised when running tests

This PR silences a bunch of warnings that get raised when running the test suite in this repo.

* Rename file

* Remove misleading assertion

It's probably best to remove this assertion because it's existence is necessitated by a hack. For Carbon to render a DataTable with expansion correctly, we need to have a placeholder DOM element to render when a table row is not in an expanded state. Previously, we used an empty `<div />`, but this would lead to a `validateDOMNesting` warning in the browser console. That's because the only permitted children for a `<tbody>` tag are zero or more `<tr>`s. So, I chose instead to a render a table row but then to hide it from the UI using CSS. This works fine in the browser but obviously won't in a testing environment. So I've elected to remove the assertion altogether.