(feat) Better dashboard titles in breadcrumbs menu (#560)
This improves the display of dashboard titles in the breadcrumbs menu. More specifically, it improves how dashboard titles containing more than one word are displayed. It proposes a convention where the dashboard `name` for dashboards with multi-word titles is an underscore delimited string (e.g. `clinical_views`). Then in the title property of the `registerBreadcrumbs` invocation, we replace the underscores from the string with spaces and capitalize the result. With this, we end up with `Vitals and biometrics dashboard` instead of the previous `Vitalsandbiometrics Dashboard`.