O3-1276: Users need a way to exit the workspace when forms don't load (#683)
* O3-1276: Users need a way to exit the workspace when forms don't load

Presently, the form engine handles rendering a form and its action buttons when a user launches a clinical form. If there is a problem loading a form, its action buttons won't get loaded either (see the attached screenshot). In such a circumstance, the user cannot exit from the workspace. This situation is often very frustrating.

Ciarán has helpfully crafted up an error state design which affords the user with a way to back out of the workspace when a form fails to load. The user can click on the `Close this panel` button to exit the workspace. In future, an affordance would navigate the user to the forms list from where they could choose an alternative form to open.

This PR implements the empty state design in the workspace.

* Remove unused style declaration

* Switch stylesheet to SASS to reuse color variables