(chore) Limit workflow task parallelization (#688)
This PR limits the number of tasks that run in parallel in our CI workflow. This is so as to reduce the likelihood of the CI workflow execution running out of memory. It does so by separating the `lint` and `typescript` jobs into their own pipeline that is separate from the `build` job.