(feat) Various fixes and mods to support the migration
This commit features a combination of fixes and mods required as part of the migration to Carbon v11, React 18 and React Router v6. Additionally, it includes a bunch of UI enhancements that I deemed necessary after encountering some problems with the UI while effecting the migration. Apologies for the massive diff, there was a lot to get through.