(feat) Notify the user when programs have not been configured
This PR modifies the Programs Form logic so that an error notification is displayed when programs have not been configured in the system. This communicates clearly to the user that they cannot enrol a patient when none are configured. Further, this PR also tweaks the appearance of some UI elements in tablet mode, replacing the `light` prop (from Carbon v10) with the new `Layer` component.