(fix) Fix vitals header flagging logic (#789)
This PR fixes the logic used to determine whether to show a warning background in the vitals header. Presently, the logic fails because of a bug in the `assessAllValues` function where the key-value pairs that get iterated over do not get properly destructured from the source `Object.entries` array. I've replaced this function entirely with a function that only looks at the interpretation results for each vital sign and returns a boolean depending on whether any vital signs with abnormal interpretations exist.