(feat) Automatically revalidate orders data
This PR is a follow-up to #774. It adds a call to SWR's `mutate` function to the orders form submission handler so that when the form gets submitted successfully, a revalidation request gets broadcasted to SWR hooks using the same key. This means that the user does not need to reload the patient chart to see newly created orders.

Additionally, this PR also wraps an empty state `Tile` in the OrderBasketItemList component with the `Layer` component. This brings its look and feel in line with the rest of the app.