(chore) Change bundle size workflow event trigger (#838)
Shot in the dark trying to figure out why the compressed-size-action is not computing bundle size changes between pull requests. From my sleuthing, it appears as though the action is not correctly switching between the PR and the base branches. So it's basically ending up comparing built artefacts from the same branch 🤦🏼