Update form-entry-interop.ts
I can identify in this code that the launchFormEntry function is used inside launchFormEntryOrHtmlForms function, but it is not imported from the correct module.
 import { launchFormEntry } from '@openmrs/esm-form-entry';
This will import the correct launchFormEntry function from the @openmrs/esm-form-entry module.

Additionally, it seems like openmrsBase is not defined anywhere in this code. If openmrsBase is a global variable, it should be defined before this code is executed. Otherwise, it should be passed as a parameter to the launchFormEntryOrHtmlForms function.