(chore) Update E2E GitHub Action Job Name for Improved Clarity (#1345)
This update changes the job name of the E2E GitHub action to a more meaningful and recognizable name. This will help us easily identify the action in GitHub settings, such as in the branch protection rules page. The change will improve our workflow and make it easier for team members to navigate and manage our GitHub actions.