(feat) Comboboxes in the order form should filter out non-matching items
By default, the Carbon Combobox component does not filter out items that do not match the input string. To make this behaviour possible,
we need to pass in our own filtering function to the shouldFilterItem prop. This PR adds this functionality to several comboboxes in the
order form so that filtering happens more intuitively.