(fix) O3-3982: Validate against starting a visit at a future time (#2013)
* (fix) O3-3982: Validate against starting a visit at a future time

This PR adds validation logic that ensures a visit cannot be started at a time in the future. It does so by adding a refinement to the visitFormSchema zod schema that compares the value of the visitStartTime field with the current timestamp. If the visitStartTime is in the future, a validation error message will be shown under the field, and the form will not be submitted. The error message shown reads "Visit start time cannot be in the future".

Additionally, this PR refactors the zod schema, centralizing some of the validation logic into reusable helper functions that are hopefully more easier to maintain. It also adds some test coverage for the new validation logic.

* Review feedback

---------

Co-authored-by: Dennis Kigen <kigen.work@gmail.com>