(fix) O3-4039: Enable backdating a prescription's start date
This PR fixes an issue where the Start date field of the drug order form was not being set correctly. The default value for the field was being set to the current date. Instead, the field should be set to the value of the medication order's `dateActivated` property (when editing an existing order) or the current date (when creating a new order).

This PR also refactors the MedicationsDetailsTable component to leverage the new MedicationsDisplaySection sub-component. The idea with this refactor is to clean up the logic for rendering the medications section and to make the component more modular and easier to maintain.

Ticket: https://openmrs.atlassian.net/browse/O3-4039