(fix) Fix `Last completed date` in forms dashboard (#2059)
This PR fixes an issue where the `Last completed date` in the forms dashboard was not being displayed correctly. The logic that determines the last completed date was not correctly sorting the encounters by date, and was not using the most recent encounter date. This PR fixes these issues and ensures that the last completed date is the most recent encounter date for each form.