(fix) Conditions list not updating after deleting a condition
Deleting a condition from the conditions overview tile in the Patient Summary page does not update the conditions list. Whilst the mutate function is getting called, the actual revalidation request is not being made. Because the `useConditions` hook takes in a `patientUuid` parameter, SWR uses that to invalidate the cache. The underlying issue here is that the `ConditionsActionMenu` component invoked in the `ConditionsOverview` component does not pass in the `patientUuid` parameter. This PR fixes that and gets the conditions list to update after deleting a condition.