(feat) order-basket: Improve a11y and styling of status tags (#2152)
This change improves the styling of status tags in the Order basket so that they are more consistent with the [designs](https://zeroheight.com/23a080e38/p/030b9a-order-item).
It also adds a11y attributes to the status labels so that they are more accessible. A desired side effect of these changes is that using these a11y
attributes will make the Order basket e2e tests more robust.