(fix) visit-form: Fix visit stop datetime field validations
This PR fixes validations in the visit form's stop date and stop time field in edit mode. Presently, when the user selects a stop date, that value does not get included in the payload sent to the backend. That's because logic exists that ensures that the stop datetime only gets included in the payload if the user selects a stop date, stop time, and stop time format. All of these fields are required to be filled out for the backend to accept the payload. This PR adds validation logic that ensures that if a stop date is selected but a stop time and stop time format are not, the user gets shown validation errors. This makes it clear to the user that they need to fill out all the required fields.