(feat) Restore refresh data button in results viewer
This PR restores a `Refresh data` button in the results viewer UI. This is an affordance for users to reload
data in the results viewer in lieu of a proper mechanism for automatically revalidating data. When clicked, this button will trigger
a page reload courtesy of the `window.location.reload()` method.