(feat) O3-4040: Enable deleting program enrollments (#2133)
(feat) O3-4040: Enable deleting program enrollments

This PR adds the ability to delete program enrollments to both the Programs data tiles. It does so by replacing the erstwhile "Edit" button with an overflow menu containing two options: "Edit" and "Delete". Clicking the "Delete" button launches a confirmation modal that prompts the user to confirm the delete action. Once a program enrollment gets deleted, a success snackbar notification confirming the action is shown on the screen.

This PR also augments existing unit and e2e tests to cover this new functionality.

Co-authored-by: Dennis Kigen <kigen.work@gmail.com>