(feat) O3-4444: Map 'Start date' field in drug order form to `dateActivated`
This PR makes the following changes to the medications app:

- Maps the `startDate` field to the `dateActivated` property in the drug order form. Previously, the payload sent to the Order resource
in the backend when creating a new drug order or editing an existing one was missing the `dateActivated` property. This meant that the backend
wound up setting the `dateActivated` property to the current date and time. This is especially problematic when editing an existing order, as the
current date was being used as the `dateActivated` property. This PR fixes this issue by mapping the `startDate` field to the `dateActivated` property and having its value
default to the current date and time for new orders, and using the existing value for existing orders.
- Amends the "Start date" field in the drug order form to read "Activation date" instead of "Start date".
- Amends the "Start date" column header in the medications details table to read "Activation date" instead of "Start date".