(fix) Remove additional top border from contact details panel (#2242)
This is a follow up fix necessitated by https://github.com/openmrs/openmrs-esm-core/pull/1284, which introduced a new top border to the contact details panel. The rationale behind that change was to ensure that the contact details panel always has a top border, and that the border should be added by the [PatientBannerContactDetails](https://github.com/openmrs/openmrs-esm-core/blob/main/packages/framework/esm-styleguide/src/patient-banner/contact-details/patient-banner-contact-details.component.tsx#L183) component in the styleguide.

Consumers of the reusable banner components from the styleguide should not need to define the top border in their styles. That should be the responsibility of the styleguide. To that end, this PR removes the additional top border from the PatientBanner component.