(chore) Run E2E tests on Firefox
Changes the browser configuration in the Playwright test setup from Chrome to Firefox. This is an attempt to try and get a handle on the failing allergies e2e test. In my local testing,
I've found that the test passes when running on Firefox. This change will help me get a better understanding of the issue.