(test) Simplify secondary diagnosis selection in visit note e2e test
This commit simplifies the secondary diagnosis selection in the visit
note e2e test by setting a diagnosis with a shorter name. Previously, we set
'Infection by Ascaris Lumbricoides' as the secondary diagnosis. This is a
fairly long name and it was causing issues with the test. More specifically, in newer versions
of Carbon, when a long name is set, the Tag component will truncate the text and display a tooltip.
The presence of the tooltip was causing the test to fail because Playwright would
encounter a strict mode violation.

By setting the secondary diagnosis to 'GI upset', we avoid this issue and the test
passes consistently.