(fix) Allergies form should not change value to undefined (#2329)
In the E2E tests, when Playwright clicks on the allergen, two __item_clicked__ state changes are fired by Downshift (the library Carbon uses to implement the ComboBox). The first time this is fired, the selectedItem property is properly set to a value, but the second time it's set to `undefined`. Here we catch any attempts to set the value to `undefined` in the ComboBox onChange() handler and don't update the form state. However, clicking the button to clear the entry appears to send `null`, so this still works. This does lose the functionality where clicking on an already-selected item deselects that item (since that sets the current value to `undefined`), but that feels like a reasonable tradeoff.