(feat) Centralize allergy form validation logic (#2333)
* (feat) Centralize allergy form validation logic

This PR centralizes validation logic in the allergies form workspace so that the
zod schema is the source of truth for validation rules. Presently, some validation logic exists in useEffect
hooks that determines whether the submit button should be disabled. This logic isn't tied to any display logic, which means
that validation errors are not shown to the user when they are encountered.

With this PR, fields are validated when the user attempts to save or update an allergy. If validation fails, the user is presented with
the appropriate error message.

* Update tests to account for new allergy form validation

This commit updates the AllergyForm test suite to account for the new allergy form validation logic. It also effects some useful refactoring to the allergy form workspace component. Notably, that includes inlining the AllergicReactionsForm component, annotating the zod schema with useful comments, and replacing enums with constants in line with best practices.