(fix) Fix nav-groups rendered in Patient Chart
Because the patient-chart uses a low-level API (the assigned extensions
for the slot derived directly from the extension store), it sometimes
happens that a nav-group is accessed here before the extension slot it
contains has been registered. Here we register the slot and then re-render
the chart component to ensure that the extensions attached to that slot
are also available.